requirements.txt
Django==2.0
django-taggit==1.2.0
Markdown==3.1.1
mysqlclient==1.4.6
pytz==2019.3
